11.06.2023

+++ Artist Hub/myapp/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Contact(models.Model):
 	name=models.CharField(max_length=100)
 	email=models.EmailField()
-	mobile=models.PositiveSmallIntegerField()
+	mobile=models.BigIntegerField()
 	message=models.TextField(max_length=500)
 
 	def __str__(self):
